api-login/requirements.txt
fastapi
sqlalchemy
asyncpg
werkzeug
pyjwt
uvicorn
